+++ dolgobot/src/com/redeyesgang/DB/IWorkWithUsers.java
@@ -3,10 +3,10 @@ package com.redeyesgang.DB;
 import java.sql.SQLException;
 
 public interface IWorkWithUsers {
-    void createUser(int telegramUid, String firstName,String dolgobotLogin) throws SQLException, OnCreateException;
-    void createUser(int telegramUid, String firstName,String dolgobotLogin,String lastName) throws SQLException, OnCreateException;
-    void createGroup(int telegramUid, String groupName) throws SQLException, OnCreateException;
-    void deleteGroup(int telegramUid, String groupName) throws SQLException, OnCreateException;
-    void addUserToGroup(int telegramUid,String groupName) throws SQLException, OnCreateException;
-    void deleteUserFromGroup(int telegramUid, String groupName) throws SQLException, OnCreateException;
+    void createUser(long telegramUid, String firstName,String dolgobotLogin) throws SQLException, OnCreateException;
+    void createUser(long telegramUid, String firstName,String dolgobotLogin,String lastName) throws SQLException, OnCreateException;
+    void createGroup(long telegramUid, String groupName) throws SQLException, OnCreateException;
+    void deleteGroup(long telegramUid, String groupName) throws SQLException, OnCreateException;
+    void addUserToGroup(long telegramUid,String groupName) throws SQLException, OnCreateException;
+    void deleteUserFromGroup(long telegramUid, String groupName) throws SQLException, OnCreateException;
 }
