+++ dolgobot/src/com/redeyesgang/DB/WorkWithUsers.java
@@ -12,7 +12,7 @@ public class WorkWithUsers implements IWorkWithUsers {
     }
 
     @Override
-    public void createUser(int telegramUid, String firstName, String dolgobotLogin) throws SQLException, OnCreateException {
+    public void createUser(long telegramUid, String firstName, String dolgobotLogin) throws SQLException, OnCreateException {
         PreparedStatement ps = _con.prepareStatement(_props.getProperty("getUser"));
         ps.setString(1,dolgobotLogin);
         ResultSet rs = ps.executeQuery();
@@ -24,7 +24,7 @@ public class WorkWithUsers implements IWorkWithUsers {
         rs.close();
         ps.close();
         ps = _con.prepareStatement(_props.getProperty("createUser"));
-        ps.setInt(1,telegramUid);
+        ps.setLong(1,telegramUid);
         ps.setString(2,firstName);
         ps.setString(3,dolgobotLogin);
         ps.executeUpdate();
@@ -32,7 +32,7 @@ public class WorkWithUsers implements IWorkWithUsers {
     }
 
     @Override
-    public void createUser(int telegramUid, String firstName, String dolgobotLogin, String lastName) throws SQLException, OnCreateException {
+    public void createUser(long telegramUid, String firstName, String dolgobotLogin, String lastName) throws SQLException, OnCreateException {
         PreparedStatement ps = _con.prepareStatement(_props.getProperty("getUser"));
         ps.setString(1,dolgobotLogin);
         ResultSet rs = ps.executeQuery();
@@ -44,7 +44,7 @@ public class WorkWithUsers implements IWorkWithUsers {
         rs.close();
         ps.close();
         ps = _con.prepareStatement(_props.getProperty("createUserWithSN"));
-        ps.setInt(1,telegramUid);
+        ps.setLong(1,telegramUid);
         ps.setString(2,firstName);
         ps.setString(3,lastName);
         ps.setString(4,dolgobotLogin);
@@ -53,7 +53,7 @@ public class WorkWithUsers implements IWorkWithUsers {
     }
 
     @Override
-    public void createGroup(int telegramUid, String groupName) throws SQLException, OnCreateException {
+    public void createGroup(long telegramUid, String groupName) throws SQLException, OnCreateException {
         PreparedStatement ps = _con.prepareStatement(_props.getProperty("getGroupID"));
         ps.setString(1,"groupName");
         ResultSet rs = ps.executeQuery();
@@ -67,7 +67,7 @@ public class WorkWithUsers implements IWorkWithUsers {
 
         ps = _con.prepareStatement(_props.getProperty("createGroup"));
         ps.setString(1,groupName);
-        ps.setInt(2, telegramUid);
+        ps.setLong(2, telegramUid);
         ps.executeUpdate();
         ps.close();
 
@@ -75,7 +75,7 @@ public class WorkWithUsers implements IWorkWithUsers {
         stmt.execute(QueryBuilderForGroup.getCreateGroupQuery(groupName));
         stmt.close();
         ps = _con.prepareStatement(QueryBuilderForGroup.getInsertToGroupQuery(groupName));
-        ps.setInt(1,telegramUid);
+        ps.setLong(1,telegramUid);
         ps.executeUpdate();
         ps.close();
 
@@ -88,20 +88,20 @@ public class WorkWithUsers implements IWorkWithUsers {
         rs.close();
 
         ps = _con.prepareStatement(_props.getProperty("insertUiG"));
-        ps.setInt(1,telegramUid);
+        ps.setLong(1,telegramUid);
         ps.setInt(2,groupID);
         ps.executeUpdate();
         ps.close();
     }
 
     @Override
-    public void deleteGroup(int telegramUid, String groupName) throws SQLException, OnCreateException {
+    public void deleteGroup(long telegramUid, String groupName) throws SQLException, OnCreateException {
         PreparedStatement ps = _con.prepareStatement(_props.getProperty("getAdmin"));
         ps.setString(1,groupName);
         ResultSet rs = ps.executeQuery();
-        int adminID;
+        long adminID;
         if (rs.next()) {
-               adminID = rs.getInt(1);
+               adminID = rs.getLong(1);
         }
         else {
             rs.close();
@@ -117,7 +117,7 @@ public class WorkWithUsers implements IWorkWithUsers {
     }
 
     @Override
-    public void addUserToGroup(int telegramUid, String groupName) throws SQLException, OnCreateException {
+    public void addUserToGroup(long telegramUid, String groupName) throws SQLException, OnCreateException {
         int groupID;
         PreparedStatement ps = _con.prepareStatement(_props.getProperty("getGroupID"));
         ps.setString(1,groupName);
@@ -131,20 +131,20 @@ public class WorkWithUsers implements IWorkWithUsers {
         ps.close();
 
         ps = _con.prepareStatement(QueryBuilderForGroup.getInsertToGroupQuery(groupName));
-        ps.setInt(1,telegramUid);
+        ps.setLong(1,telegramUid);
         ps.executeUpdate();
         ps.close();
 
 
         ps = _con.prepareStatement(_props.getProperty("insertUiG"));
-        ps.setInt(1,telegramUid);
+        ps.setLong(1,telegramUid);
         ps.setInt(2,groupID);
         ps.executeUpdate();
         ps.close();
     }
 
     @Override
-    public void deleteUserFromGroup(int telegramUid, String groupName) throws SQLException, OnCreateException {
+    public void deleteUserFromGroup(long telegramUid, String groupName) throws SQLException, OnCreateException {
         int groupID;
         PreparedStatement ps = _con.prepareStatement(_props.getProperty("getGroupID"));
         ps.setString(1,groupName);
@@ -156,20 +156,20 @@ public class WorkWithUsers implements IWorkWithUsers {
         }
         rs.close();
         ps.close();
-
+        // TO DO: delete from group!
         ps = _con.prepareStatement(QueryBuilderForGroup.getDeleteGroupQuery(groupName));
         ps.executeUpdate();
         ps.close();
     }
 
-    public int getGroupAdminID(String groupName) throws SQLException, OnCreateException {
+    public long getGroupAdminID(String groupName) throws SQLException, OnCreateException {
         PreparedStatement ps  = _con.prepareStatement(_props.getProperty("getAdmin"));
         ps.setString(1,groupName);
         ResultSet rs = ps.executeQuery();
-        int res;
+        long res;
 
         if (rs.next()) {
-            res = rs.getInt(1);
+            res = rs.getLong(1);
         }else {
             throw new OnCreateException("Не сущесвует группы с таким именем!");
         }
